feat: final improvements
- Re-organized the files to be consistent between features
- Removed unused files
- Moved flow types to types folder
