+++ packages/ui/src/components/ui/input.tsx
@@ -1,4 +1,4 @@
-import { forwardRef, ComponentProps, InputHTMLAttributes } from 'react'
+import { forwardRef, InputHTMLAttributes } from 'react'
 
 import { cn } from '@/lib/utils'
 import { cva, type VariantProps } from 'class-variance-authority'
