+++ packages/workbench/index.tsx
@@ -1,11 +1,11 @@
 export { EventNode } from './src/publicComponents/event-node'
 export { ApiNode } from './src/publicComponents/api-node'
 export { NoopNode } from './src/publicComponents/noop-node'
-export { BaseNode } from './src/publicComponents/base-node'
-export { BaseHandle } from './src/publicComponents/base-handle'
+export { BaseNode } from './src/publicComponents/base-node/base-node'
+export { BaseHandle } from './src/publicComponents/base-node/base-handle'
 
 export { Position } from '@xyflow/react'
-export type { EventNodeData, ApiNodeData } from './src/views/flow/nodes/nodes.types'
+export type { EventNodeData, ApiNodeData } from './src/types/flow'
 export type {
   ApiNodeProps,
   BaseNodeProps,
@@ -13,4 +13,3 @@ export type {
   EventNodeProps,
   NoopNodeProps,
 } from './src/publicComponents/node-props'
-export { Button } from './src/components/ui/button'
