+++ packages/workbench/src/components/endpoints/endpoint-call.tsx
@@ -1,11 +1,11 @@
 import { Sidebar } from '@/components/sidebar/sidebar'
+import { ApiEndpoint } from '@/types/endpoint'
 import { Button, Input, Panel } from '@motiadev/ui'
 import { Loader2, Play, X } from 'lucide-react'
 import { FC, useEffect, useMemo, useState } from 'react'
 import { EndpointBadge } from './endpoint-badge'
 import { EndpointResponse } from './endpoint-response'
 import { EndpointResponseSchema } from './endpoint-response-schema'
-import { ApiEndpoint } from './hooks/use-get-endpoints'
 import { useJsonSchemaToJson } from './hooks/use-json-schema-to-json'
 import { usePathParams } from './hooks/use-path-params'
 import { JsonEditor } from './json-editor'
