+++ packages/workbench/src/components/endpoints/endpoint-response.tsx
@@ -1,9 +1,9 @@
+import { useStateStream } from '@/components/endpoints/hooks/use-state-stream'
 import { useThemeStore } from '@/stores/use-theme-store'
 import { Panel } from '@motiadev/ui'
 import { XCircle } from 'lucide-react'
 import { FC, useMemo } from 'react'
 import ReactJson from 'react18-json-view'
-import { useStateStream } from '@/components/endpoints/hooks/use-state-stream'
 
 type EndpointResponseProps = {
   responseCode: number | string | undefined
