+++ packages/workbench/src/components/endpoints/hooks/use-get-endpoints.ts
@@ -1,17 +1,5 @@
 import { useStreamGroup } from '@motiadev/stream-client-react'
-
-type ApiRouteMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH' | 'OPTIONS' | 'HEAD'
-type QueryParam = { name: string; description: string }
-
-export type ApiEndpoint = {
-  id: string
-  method: ApiRouteMethod
-  path: string
-  description?: string
-  queryParams?: QueryParam[]
-  responseSchema?: Record<string, any> // eslint-disable-line @typescript-eslint/no-explicit-any
-  bodySchema?: Record<string, Record<string, any>> // eslint-disable-line @typescript-eslint/no-explicit-any
-}
+import { ApiEndpoint } from '@/types/endpoint'
 
 export const useGetEndpoints = () => {
   const { data: endpoints } = useStreamGroup<ApiEndpoint>({
