+++ packages/workbench/src/components/flow/node-organizer.tsx
@@ -1,7 +1,7 @@
-import { Edge, Node } from '@xyflow/react'
+import { ApiNodeData, EdgeData, EventNodeData } from '@/types/flow'
+import { Edge, Node, useNodesInitialized, useReactFlow } from '@xyflow/react'
 import dagre from 'dagre'
-import { useEffect, useRef } from 'react'
-import { EventNodeData, EdgeData, ApiNodeData } from '../nodes/nodes.types'
+import React, { useEffect, useRef } from 'react'
 
 const organizeNodes = (
   nodes: Node<EventNodeData | ApiNodeData>[],
@@ -10,7 +10,6 @@ const organizeNodes = (
   const dagreGraph = new dagre.graphlib.Graph({ compound: true })
   dagreGraph.setDefaultEdgeLabel(() => ({}))
 
-  // Top-to-bottom layout
   dagreGraph.setGraph({ rankdir: 'TB', ranksep: 80, nodesep: 60, edgesep: 20, ranker: 'tight-tree' })
 
   nodes.forEach((node) => {
@@ -33,6 +32,10 @@ const organizeNodes = (
   dagre.layout(dagreGraph)
 
   return nodes.map((node) => {
+    if (node.position.x !== 0 || node.position.y !== 0) {
+      return node
+    }
+
     const { x, y } = dagreGraph.node(node.id)
     const position = {
       x: x - (node.measured?.width ?? 0) / 2,
@@ -43,18 +46,31 @@ const organizeNodes = (
   })
 }
 
-export const useOrganizeNodes = (
-  nodes: Node<EventNodeData | ApiNodeData>[],
-  edges: Edge<EdgeData>[],
-  setNodes: (nodes: Node<EventNodeData | ApiNodeData>[]) => void,
-) => {
-  const organizedRef = useRef<boolean>(false)
+type Props = {
+  onInitialized: () => void
+}
+
+export const NodeOrganizer: React.FC<Props> = ({ onInitialized }) => {
+  const { setNodes, getNodes, getEdges, fitView } = useReactFlow()
+  const nodesInitialized = useNodesInitialized()
+  const initialized = useRef(false)
 
   useEffect(() => {
-    if (!nodes.length || !edges.length || !nodes[0].measured || organizedRef.current) return
+    if (nodesInitialized && !initialized.current) {
+      initialized.current = true
 
-    const layoutedNodes = organizeNodes(nodes, edges)
-    setNodes(layoutedNodes)
-    organizedRef.current = true
-  }, [nodes, edges, setNodes])
+      const nodes = getNodes() as Node<EventNodeData | ApiNodeData>[]
+      const edges = getEdges() as Edge<EdgeData>[]
+      const organizedNodes = organizeNodes(nodes, edges)
+
+      setNodes(organizedNodes)
+      onInitialized()
+
+      setTimeout(async () => {
+        await fitView()
+      }, 1)
+    }
+  }, [nodesInitialized, onInitialized, setNodes, getNodes, getEdges, fitView])
+
+  return null
 }
