+++ packages/workbench/src/components/flow/nodes/noop-flow-node.tsx
