+++ packages/workbench/src/components/observability/events/trace-event.tsx
@@ -1,5 +1,5 @@
+import { TraceEvent as TraceEventType } from '@/types/observability'
 import React, { memo } from 'react'
-import { TraceEvent as TraceEventType } from '../../../types/observability'
 import { TraceEmitEvent } from './trace-emit-event'
 import { TraceLogEvent } from './trace-log-event'
 import { TraceStateEvent } from './trace-state-event'
