--- packages/workbench/src/publicComponents/colorMap.ts
@@ -1,6 +1,0 @@
-export const colorMap = {
-  event: 'rgb(0,117,255)',
-  api: 'rgb(189,255,0)',
-  noop: 'rgb(255,49,234)',
-  cron: 'rgb(255,113,11)',
-}
