+++ packages/workbench/src/publicComponents/node-props.tsx
@@ -1,4 +1,4 @@
-import { ApiNodeData, CronNodeData, EventNodeData, NoopNodeData } from '../views/flow/nodes/nodes.types'
+import { ApiNodeData, CronNodeData, EventNodeData, NoopNodeData } from '../types/flow'
 
 export type NodeProps = EventNodeProps | NoopNodeProps | ApiNodeProps | CronNodeProps
 export type BaseNodeProps = {
