+++ playground/package.json
@@ -16,6 +16,7 @@
     "lint:tsx": "eslint --config ../eslint.config.tsx.js"
   },
   "dependencies": {
+    "@motiadev/ui": "workspace:*",
     "@motiadev/core": "workspace:*",
     "@motiadev/workbench": "workspace:*",
     "motia": "workspace:*",
