+++ playground/steps/default-node/00-noop.step.tsx
@@ -1,5 +1,6 @@
 import React from 'react'
-import { BaseNode, Button, NoopNodeProps } from '@motiadev/workbench'
+import { BaseNode, NoopNodeProps } from '@motiadev/workbench'
+import { Button } from '@motiadev/ui'
 
 /**
  * For more information on how to override UI nodes, check documentation https://www.motia.dev/docs/workbench/ui-steps
