+++ playground/steps/parallelMergeState/noopStartButton.step.tsx
@@ -1,4 +1,5 @@
-import { BaseNode, Button, NoopNodeProps } from '@motiadev/workbench'
+import { BaseNode, NoopNodeProps } from '@motiadev/workbench'
+import { Button } from '@motiadev/ui'
 import React from 'react'
 
 export const Node: React.FC<NoopNodeProps> = (data) => {
